feat: Add 'preventEmpty' option to InputToggleArgs
This commit adds a new 'preventEmpty' option to the InputToggleArgs interface in the _types.tsx file. This option allows preventing empty values in the InputToggle component. The default value is set to true.

The change also includes modifications in the index.stories.tsx and index.tsx files to support the new option. The PreventEmpty story is added to demonstrate the usage of the 'preventEmpty' option.